+++ src/main/frontend/src/member/myPage/LikeList.jsx
@@ -6,6 +6,7 @@ import MyContent from "./MyContent";
 import MyQNA from "./MyQNA";
 import MyComment from "./MyComment";
 import {fetchData} from "../../common/NovelDetailFetch2";
+import LikeListItem from "./LikeListItem";
 
 function LikeList(props) {
     const navi = useNavigate();
@@ -19,10 +20,7 @@ function LikeList(props) {
     const [pages, setPages] = useState([1]);
     const [likeList, setLikeList] = useState([{}]);
     const [adult, setAdult] = useState([""])
-    const adultsOnly = {
-        Y: true,
-        N: false
-    };
+
 
     useEffect(() => {
         setNowPage(props.defaultPage);
@@ -33,7 +31,7 @@ function LikeList(props) {
     }, [props.data, nowPage]);
 
     const requestData = () => {
-        axios.get(`/myPage/likeList?id=${sessionStorage.getItem('id')}&page=${nowPage}&size=10`)
+        axios.get(`/myPage/likeList?id=${sessionStorage.getItem('id')}`)
             .then(res => {
 
                 console.log(res.data);
@@ -56,37 +54,6 @@ function LikeList(props) {
                 alert(`통신에 실패했습니다. err : ${err}`);
             })
     }
-
-    // const likeClickHandler = async (like) => {
-    //     axios.put('/novelDetailLike', null, {
-    //         params: {
-    //             novelIdx: like.novelIdx,
-    //             id: sessionStorage.getItem('id')
-    //         }
-    //     })
-    //         .then((res) => {
-    //             console.log(res);
-    //             // console.log(res);
-    //             axios.get("/novelDetail", {
-    //                 params: {
-    //                     title: like.novelTitle,
-    //                     ebookCheck: like.ebookCheck
-    //                 }
-    //             })
-    //                 .then(res2 => {
-    //                     console.log(res2);
-    //                     setNovelLikeList(res2.data.novelLikeList);
-    //                     setNovelLikeCount(res2.data.novelLikeCount);
-    //                 })
-    //                 .catch(err => {
-    //                     console.log(err.message)
-    //                 })
-    //
-    //         })
-    //         .catch(err => {
-    //             console.log(err.message);
-    //         })
-    // }
 
     const handleLinkClick = async (like) => {
         console.log(like);
@@ -112,64 +79,10 @@ function LikeList(props) {
             <Row>
                 {
                     likeList.length != 0
-                        ? likeList.map((like, index) => {
+                        ? likeList.map(like => {
                             return (
-                                <Col className={"rank-item"} sm={2} key={index}>
-
-                                    <div>
-                                        <div className={"rank-item-img"}>
-                                            {/*세션 영역에 저장된 성인 여부에 따라 이미지 보이는 거 다르게 해야함 */}
-                                            <Link
-                                                onClick={e => {
-                                                    e.preventDefault();
-                                                    handleLinkClick(like);
-                                                }} to={`/novelDetail/${like.title}`}>
-                                                {
-                                                    adultsOnly[adult] ? <img
-                                                            src={"https://page.kakaocdn.net/pageweb/2.12.2/public/images/img_age_19_Thumbnail_43.svg"}
-                                                            alt={"Adults content"}/>
-                                                        : <img src={like.novelThumbnail} alt=""/>
-                                                }
-                                            </Link>
-                                            <button className={'rank-num btn align-items-end'} onClick={() => {
-                                                axios.put('/novelDetailLike', null, {
-                                                    params: {
-                                                        novelIdx: like.novelIdx,
-                                                        novelTitle: like.novelTitle,
-                                                        ebookCheck: like.ebookCheck,
-                                                        id: sessionStorage.getItem('id')
-                                                    }
-                                                })
-                                                    .then((res) => {
-                                                        console.log(res);
-                                                        // console.log(res);
-                                                        axios.get("/novelDetail", {
-                                                            params: {
-                                                                title: like.novelTitle,
-                                                                ebookCheck: like.ebookCheck
-                                                            }
-                                                        })
-                                                            .then(res2 => {
-                                                                console.log(res2);
-                                                                setNovelLikeList(res2.data.novelLikeList);
-                                                                setNovelLikeCount(res2.data.novelLikeCount);
-                                                            })
-                                                            .catch(err => {
-                                                                console.log(err.message)
-                                                            })
-
-                                                    })
-                                                    .catch(err => {
-                                                        console.log(err.message);
-                                                    })
-                                            }}>
-                                                <i className="bi bi-heart-fill text-purple"></i>
-                                            </button>
-                                        </div>
-                                    </div>
-                                    <div className={"rank-info w-100"}>
-                                        <p className={"item-title"}>{like.novelTitle}</p>
-                                    </div>
+                                <Col className={"rank-item"} sm key={like.novelIdx}>
+                                    <LikeListItem like={like}/>
                                 </Col>
                             )
                         })
